+++ qom/ui/axes/DynamicAxis.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.ui.axes.DynamicAxis'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-09-17'
-__updated__ = '2022-04-24'
+__updated__ = '2022-06-11'
 
 # dependencies
 import logging
@@ -23,7 +23,7 @@ class DynamicAxis(BaseAxis):
     Parameters
     ----------
     axis : str
-        Name of the axis, "X", "Y", "Z" or "V".
+        Name of the axis, or "X", "Y", "Z" or "V".
     axis_params : dict or list
         Values for the axis supporting a list of values or a dictionary containing the range of values with keys "min", "max", "dim" and "scale" or the values themselves under key "val".
     plotter_params : dict
@@ -31,7 +31,7 @@ class DynamicAxis(BaseAxis):
             ==============  ====================================================
             key             value
             ==============  ====================================================
-            "name"          (*str*) display name of the axis.
+            "name"          (*str*) display name of the axis. Default is 'DynamicAxis'.
             "unit"          (*str*) unit of the plots.
             ==============  ====================================================
     """
@@ -47,3 +47,7 @@ class DynamicAxis(BaseAxis):
 
         # set unit
         self.unit = plotter_params.get(axis.lower() + '_unit', self.axis_defaults['unit'])
+
+        # update label
+        if self.label == '' and not 'DynamicAxis' in self.name:
+            self.label = self.name if self.unit == '' else '{name} ({unit})'.format(name=self.name, unit=self.unit)